Clear remote shares between scenarios
Although groups are removed after each scenario it seems that their
remote group shares are not, so there might be dangling remote group
shares that could interfere with the following scenarios. To solve that
a new endpoint is added to the "testing" app to explicitly clear the
remote shares.

Note that any dangling remote storage was already removed with the OCC
command, so only the "share_external" table is cleared.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>